+++ paymentsheet/src/main/java/com/stripe/android/checkout/CheckoutEmbeddedConfigurationFactory.kt
@@ -1,12 +1,18 @@
 package com.stripe.android.checkout
 
 import com.stripe.android.checkout.injection.AppName
+import com.stripe.android.paymentelement.CardFundingFilteringPrivatePreview
 import com.stripe.android.paymentelement.CheckoutSessionPreview
 import com.stripe.android.paymentelement.EmbeddedPaymentElement
 import com.stripe.android.paymentsheet.repositories.CheckoutSessionResponse
 import javax.inject.Inject
+import com.stripe.android.paymentsheet.CardFundingFilteringPrivatePreview as PaymentSheetCardFundingFilteringPrivatePreview
 
-@OptIn(CheckoutSessionPreview::class)
+@OptIn(
+    CheckoutSessionPreview::class,
+    CardFundingFilteringPrivatePreview::class,
+    PaymentSheetCardFundingFilteringPrivatePreview::class,
+)
 internal class CheckoutEmbeddedConfigurationFactory @Inject constructor(
     @AppName private val appName: String,
 ) {
@@ -22,6 +28,12 @@ internal class CheckoutEmbeddedConfigurationFactory @Inject constructor(
             )
             .billingDetailsCollectionConfiguration(
                 configuration.toBillingDetailsCollectionConfiguration(checkoutSessionResponse)
+            )
+            .preferredNetworks(configuration.paymentElementConfiguration.preferredNetworks)
+            .paymentMethodOrder(configuration.paymentElementConfiguration.paymentMethodOrder)
+            .allowedCardFundingTypes(configuration.paymentElementConfiguration.allowedCardFundingTypes.asPaymentSheet())
+            .opensCardScannerAutomatically(
+                configuration.paymentElementConfiguration.opensCardScannerAutomatically
             )
             .termsDisplay(configuration.paymentElementConfiguration.termsDisplay.asPaymentSheet())
             .appearance(configuration.paymentElementConfiguration.appearance.asPaymentSheet())