+++ paymentsheet/src/main/java/com/stripe/android/elements/ece/ExpressButtonType.kt
@@ -2,12 +2,12 @@
 
 package com.stripe.android.elements.ece
 
-import com.stripe.android.checkout.GooglePayConfiguration
+import com.stripe.android.elements.ExpressCheckoutElement
 import com.stripe.android.paymentelement.CheckoutSessionPreview
 
 internal sealed class ExpressButtonType {
     data object Link : ExpressButtonType()
     data class GooglePay(
-        val googlePayConfiguration: GooglePayConfiguration.State,
+        val googlePayConfiguration: ExpressCheckoutElement.Configuration.GooglePayConfiguration.State,
     ) : ExpressButtonType()
 }